trxnslateAPI/requirements.txt
fastapi[all]
uvicorn
google-cloud-storage
tensorflow
keras
pydantic
opencv-python